use slot_info_window_map instead of slot_info_window_vec and slot_info_map

+++ src/modules/apa_function/src/perpendicular_planner/perpendicular_park_in_planner.cc
@@ -1117,7 +1117,7 @@ const uint8_t PerpendicularInPlanner::PathPlanOnce() {
     case 1:
       break;
     default:
-      for (int i = 0; i < path_seg_vec.size() - 1; ++i) {
+      for (int i = 0; i < static_cast<int>(path_seg_vec.size()) - 1; ++i) {
         if (planner_output.gear_cmd_vec[i] !=
             planner_output.gear_cmd_vec[i + 1]) {
           break;
